requirements.txt
amqp==5.1.1
AMQPStorm==2.10.6
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
crochet==2.0.0
cryptography==41.0.1
cssselect==1.2.0
Django==4.2.2
django-celery-results==2.5.1
djangorestframework==3.14.0
filelock==3.12.2
google-play-scraper==1.2.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kombu==5.3.1
lxml==4.9.2
packaging==23.1
pamqp==2.3.0
parsel==1.8.1
prompt-toolkit==3.0.38
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
queuelib==1.6.2
redis==4.5.5
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-djangoitem==1.1.1
scrapydo==0.2.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==6.0
